% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_proxy.R
\name{config_proxy_mix}
\alias{config_proxy_mix}
\title{Generating configuration information for the proxy}
\usage{
config_proxy_mix(ip = "192.168.3.15", port = "7890")
}
\arguments{
\item{ip}{address, default 192.168.3.15}

\item{port}{mix port, default 7890}
}
\value{
a character vector contains proxy for each protocols (http, https, socks5)
}
\description{
Suitable for mixed proxies, i.e. where ip and port are consistent across different protocols.
}
\examples{
config_proxy_mix(ip = "127.0.0.1", port = "7890")
}
\seealso{
Other config: 
\code{\link{config_proxy}()}
}
\concept{config}
