% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{proxy}
\alias{proxy}
\title{Set proxy}
\usage{
proxy(config_info = config_proxy_mix())
}
\arguments{
\item{config_info}{output from \code{\link[config_proxy_mix]{proxy4you}}
or \code{\link[config_proxy_mix]{proxy4you}}}
}
\value{
set proxy for http https and socks5
}
\description{
set proxy for http https and socks5
}
\examples{
\dontrun{
proxy()
proxy(
config_proxy_mix(ip = "127.0.0.1", port = "7890"))
proxy(
config_proxy(
http = "http://127.0.0.1:7890",
https = "http://127.0.0.1:7890",
socks5 = "socks5://127.0.0.1:7891"))
}
}
